+++ src/core/rebalance.spec.ts
@@ -1,14 +1,15 @@
 import { describe, it, expect } from 'vitest';
 
 import SplitPanel from './SplitPanel';
-import { rebalanceSizes, getBalancedPanelSizeArray } from './defs';
+import { rebalanceSizes, getBalancedPanelSizeArray } from './utilBalance';
 
 describe('rebalanceSizes', () => {
   it('does not change sizes when percent sum is less than 100%', () => {
     const sizes = ["20%", "20%"];
-    const result = rebalanceSizes(sizes, 100);
+    const { balancedSizes, totalRelative } = rebalanceSizes(sizes, 100);
 
-    expect(result).toEqual([
+    expect(totalRelative).toEqual(0.4);
+    expect(balancedSizes).toEqual([
       { exactValue: 20, relativeValue: 0.2, relative: true },
       { exactValue: 20, relativeValue: 0.2, relative: true },
     ]);
@@ -16,9 +17,10 @@ describe('rebalanceSizes', () => {
 
   it('will rebalance the sizes when the percent sum is greater than 100%', () => {
     const sizes = ["20%", "20%", '80%'];
-    const result = rebalanceSizes(sizes, 100);
+    const { balancedSizes, totalRelative } = rebalanceSizes(sizes, 100);
 
-    expect(result).toEqual([
+    expect(totalRelative).toBeCloseTo(1.2);
+    expect(balancedSizes).toEqual([
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: true },
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: true },
       { exactValue: 66.66666666666666, relativeValue: 0.6666666666666666, relative: true },
@@ -27,9 +29,10 @@ describe('rebalanceSizes', () => {
 
   it('will rebalance sizes when the exact sum is greater than the size', () => {
     const sizes = ["20px", "20px", '80px'];
-    const result = rebalanceSizes(sizes, 100);
+    const { balancedSizes, totalRelative } = rebalanceSizes(sizes, 100);
 
-    expect(result).toEqual([
+    expect(totalRelative).toBeCloseTo(1.2);
+    expect(balancedSizes).toEqual([
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: false },
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: false },
       { exactValue: 66.66666666666666, relativeValue: 0.6666666666666666, relative: false },
@@ -38,9 +41,10 @@ describe('rebalanceSizes', () => {
 
   it('will rebalance sizes when the percent and exact sums are greater than the size', () => {
     const sizes = ["20%", "20%", '80px'];
-    const result = rebalanceSizes(sizes, 100);
+    const { balancedSizes, totalRelative } = rebalanceSizes(sizes, 100);
 
-    expect(result).toEqual([
+    expect(totalRelative).toBeCloseTo(1.2);
+    expect(balancedSizes).toEqual([
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: true },
       { exactValue: 16.666666666666664, relativeValue: 0.16666666666666666, relative: true },
       { exactValue: 66.66666666666666, relativeValue: 0.6666666666666666, relative: false },
@@ -79,28 +83,8 @@ describe('getBalancedPanelSizeArray', () => {
 
   it.each([
     {
-      message: 'balances the sizes for four panels and one 25% size',
-      size: '25%',
-      expected: [
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-      ]
-    },
-    {
       message: 'balances the sizes for four panels and four 25% sizes',
       size: ['25%', '25%', '25%', '25%'],
-      expected: [
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-        { exactValue: 250, relativeValue: 0.25, relative: true },
-      ]
-    },
-    {
-      message: 'balances the sizes for four panels and two 25% sizes',
-      size: ['25%', '25%'],
       expected: [
         { exactValue: 250, relativeValue: 0.25, relative: true },
         { exactValue: 250, relativeValue: 0.25, relative: true },
@@ -133,9 +117,9 @@ describe('getBalancedPanelSizeArray', () => {
 
     panel.satisfyConstraints({ initialize: true });
 
-    const sizes = getBalancedPanelSizeArray(size, panel.children, BASE_WIDTH);
+    const { balancedSizes } = getBalancedPanelSizeArray(panel.children.map((item, index) => ({ item, size: size[index] })), BASE_WIDTH);
 
-    expect(sizes).toEqual(expected);
+    expect(balancedSizes).toEqual(expected);
   });
 
   it('does not let the panels be smaller than the min size', () => {
@@ -158,9 +142,14 @@ describe('getBalancedPanelSizeArray', () => {
 
     panel.satisfyConstraints({ initialize: true });
 
-    const sizes = getBalancedPanelSizeArray(['50%', '50%', '0%', '0%'], panel.children, BASE_WIDTH);
+    const { balancedSizes } = getBalancedPanelSizeArray([
+      { item: panel.children[0], size: '50%' },
+      { item: panel.children[1], size: '50%' },
+      { item: panel.children[2], size: '0%' },
+      { item: panel.children[3], size: '0%' },
+    ], BASE_WIDTH);
 
-    expect(sizes).toEqual([
+    expect(balancedSizes).toEqual([
       { exactValue: 454.5454545454545, relativeValue: 0.45454545454545453, relative: true },
       { exactValue: 454.5454545454545, relativeValue: 0.45454545454545453, relative: true },
       { exactValue: 90.9090909090909, relativeValue: 0.09090909090909091, relative: false },
